JoinAPIProject/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2==2.9.10
psycopg2-binary==2.9.10